requirements (1).txt
streamlit
requests
beautifulsoup4
langchain
openai
sentence-transformers
faiss-cpu
tiktoken